Eliminate GCC-ism
clang++
-I/home/travis/build/SpiNNakerManchester/Visualiser/spynnaker_external_device_lib/
-o build/main.o -c main.cpp
In file included from main.cpp:13:
./raster_view/RasterPlot.h:43:25: warning: in-class initializer for
static data
      member of type 'const double' is a GNU extension
[-Wgnu-static-float-init]
    const static double FRAMES_PER_SECOND = 60.0;
                        ^                   ~~~~

+++ c_based_visualiser_framework/raster_view/RasterPlot.cpp
@@ -49,7 +49,7 @@ RasterPlot::RasterPlot(
 
 void RasterPlot::main_loop() {
     startFramework(argc, argv, "Raster Plot", window_width, window_height,
-                   INIT_WINDOW_X, INIT_WINDOW_Y, FRAMES_PER_SECOND);
+                   INIT_WINDOW_X, INIT_WINDOW_Y, double(FRAMES_PER_SECOND));
 }
 
 void RasterPlot::init() {
