+++ c_based_visualiser_framework/raster_view/RasterPlot.h
@@ -40,7 +40,7 @@ private:
     const static int WINDOW_BORDER = 110;
     const static int INIT_WINDOW_WIDTH = 800;
     const static int INIT_WINDOW_HEIGHT = 600;
-    const static double FRAMES_PER_SECOND = 60.0;
+    const static int FRAMES_PER_SECOND = 60;
     const static int INIT_WINDOW_X = 100;
     const static int INIT_WINDOW_Y = 100;
